Add tests for MRP authentication
The MRP server auth code now resides with the tests. Proxy is updated.
Also, the MRP fake device has support for authentication and pairing.
Some tests have been added and atvremote tests fixed.
